moved scanRequest mapper from service to seperate mapper
