+++ src/main/java/com/jasmi/xss_scanner/services/ScanRequestService.java
@@ -3,6 +3,7 @@ package com.jasmi.xss_scanner.services;
 import com.jasmi.xss_scanner.dtos.ScanRequestInputDto;
 import com.jasmi.xss_scanner.dtos.ScanRequestOutputDto;
 import com.jasmi.xss_scanner.exceptions.RecordNotFoundException;
+import com.jasmi.xss_scanner.mappers.ScanRequestMapper;
 import com.jasmi.xss_scanner.models.ScanRequest;
 import com.jasmi.xss_scanner.repositories.ScanRequestRepository;
 import org.springframework.stereotype.Service;
@@ -14,15 +15,17 @@ import java.util.stream.Collectors;
 @Service
 public class ScanRequestService {
     private final ScanRequestRepository scanRequestRepository;
+    private final ScanRequestMapper scanRequestMapper;
 
-    public ScanRequestService(ScanRequestRepository scanRequestRepository) {
+    public ScanRequestService(ScanRequestRepository scanRequestRepository, ScanRequestMapper scanRequestMapper) {
         this.scanRequestRepository = scanRequestRepository;
+        this.scanRequestMapper = scanRequestMapper;
     }
 
     public List<ScanRequestOutputDto> getAllScanRequests(){
         return scanRequestRepository.findAll()
                 .stream()
-                .map((scanRequest)-> toScanRequestDto(scanRequest))
+                .map((scanRequest)-> scanRequestMapper.toScanRequestDto(scanRequest))
                 .collect(Collectors.toList());
     }
 
@@ -30,7 +33,7 @@ public class ScanRequestService {
         Optional<ScanRequest> optionalScanRequest = scanRequestRepository.findById(id);
         if(optionalScanRequest.isPresent()){
             ScanRequest scanrequest = optionalScanRequest.get();
-            return toScanRequestDto(scanrequest);
+            return scanRequestMapper.toScanRequestDto(scanrequest);
         }
         else{
             throw new RecordNotFoundException(id+" not found");
@@ -38,32 +41,10 @@ public class ScanRequestService {
     }
 
     public ScanRequestOutputDto saveScanRequest(ScanRequestInputDto scanRequest) {
-        ScanRequest sr = toScanRequest(scanRequest);
+        ScanRequest sr = scanRequestMapper.toScanRequest(scanRequest);
         ScanRequest savedScanRequest = scanRequestRepository.save(sr);
-        return toScanRequestDto(savedScanRequest);
+        return scanRequestMapper.toScanRequestDto(savedScanRequest);
 
-    }
-
-    //DTO
-    public static ScanRequestOutputDto toScanRequestDto(ScanRequest scanRequest){
-        var dto = new ScanRequestOutputDto();
-
-        dto.setId(scanRequest.getId());
-        dto.setUrl(scanRequest.getUrl());
-        dto.setRequestTimestamp(scanRequest.getRequestTimestamp());
-        dto.setUrl(scanRequest.getUrl());
-
-        return dto;
-    }
-
-    public ScanRequest toScanRequest(ScanRequestInputDto dto){
-        var scanRequest = new ScanRequest();
-
-        scanRequest.setUrl(dto.getUrl());
-        scanRequest.setRequestTimestamp(dto.getRequestTimestamp());
-        scanRequest.setImage(dto.getImage());
-
-        return scanRequest;
     }
 
 }
